+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/boardreply/BoardReplyUpdateController.java
@@ -1,6 +1,6 @@
 package ProjectDoge.StudentSoup.controller.boardreply;
 
-import ProjectDoge.StudentSoup.dto.boardreview.BoardReplyUpdateDto;
+import ProjectDoge.StudentSoup.dto.boardreply.BoardReplyUpdateDto;
 import ProjectDoge.StudentSoup.service.boardreply.BoardReplyEditService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
