+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/dto/boardreply/BoardReplyReqDto.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.dto.boardreview;
+package ProjectDoge.StudentSoup.dto.boardreply;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -10,16 +10,12 @@ public class BoardReplyReqDto {
     private Long boardId;
     private Long memberId;
     private String content;
-    private int seq;
-    private int depth;
     private int level;
 
-    public BoardReplyReqDto createBoardReply(Long boardId, Long memberId, String content, int seq, int depth, int level) {
+    public BoardReplyReqDto createBoardReply(Long boardId, Long memberId, String content, int level) {
         this.boardId = boardId;
         this.memberId = memberId;
         this.content = content;
-        this.seq = seq;
-        this.depth = depth;
         this.level = level;
 
         return this;
