+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/dto/boardreply/BoardReplyUpdateDto.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.dto.boardreview;
+package ProjectDoge.StudentSoup.dto.boardreply;
 
 
 import ProjectDoge.StudentSoup.entity.board.BoardReply;
