+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/exception/boardreply/BoardReplyContentOutOfRangeException.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.exception.boardreview;
+package ProjectDoge.StudentSoup.exception.boardreply;
 
 public class BoardReplyContentOutOfRangeException extends RuntimeException {
     public BoardReplyContentOutOfRangeException() {
