+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/exception/boardreply/BoardReplyNotOwnException.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.exception.boardreview;
+package ProjectDoge.StudentSoup.exception.boardreply;
 
 public class BoardReplyNotOwnException extends RuntimeException{
