+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/boardreply/BoardReplyDeleteService.java
@@ -2,7 +2,7 @@ package ProjectDoge.StudentSoup.service.boardreply;
 
 
 import ProjectDoge.StudentSoup.entity.board.BoardReply;
-import ProjectDoge.StudentSoup.exception.boardreview.BoardReplyNotOwnException;
+import ProjectDoge.StudentSoup.exception.boardreply.BoardReplyNotOwnException;
 import ProjectDoge.StudentSoup.repository.boardreply.BoardReplyRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
