+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/boardreply/BoardReplyFindService.java
@@ -2,8 +2,8 @@ package ProjectDoge.StudentSoup.service.boardreply;
 
 
 import ProjectDoge.StudentSoup.entity.board.BoardReply;
-import ProjectDoge.StudentSoup.exception.boardreview.BoardReplyIdNotSentException;
-import ProjectDoge.StudentSoup.exception.boardreview.BoardReplyNotFoundException;
+import ProjectDoge.StudentSoup.exception.boardreply.BoardReplyIdNotSentException;
+import ProjectDoge.StudentSoup.exception.boardreply.BoardReplyNotFoundException;
 import ProjectDoge.StudentSoup.repository.boardreply.BoardReplyRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
