+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/boardreply/BoardReplyLikeService.java
@@ -1,7 +1,7 @@
 package ProjectDoge.StudentSoup.service.boardreply;
 
 import ProjectDoge.StudentSoup.commonmodule.ConstField;
-import ProjectDoge.StudentSoup.dto.boardreview.BoardReplyDto;
+import ProjectDoge.StudentSoup.dto.boardreply.BoardReplyDto;
 import ProjectDoge.StudentSoup.entity.board.BoardReply;
 import ProjectDoge.StudentSoup.entity.board.BoardReplyLike;
 import ProjectDoge.StudentSoup.entity.member.Member;
