+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/boardreply/BoardReplyValidationService.java
@@ -1,6 +1,6 @@
 package ProjectDoge.StudentSoup.service.boardreply;
 
-import ProjectDoge.StudentSoup.exception.boardreview.BoardReplyContentOutOfRangeException;
+import ProjectDoge.StudentSoup.exception.boardreply.BoardReplyContentOutOfRangeException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
